requirements.txt
antlr4-python3-runtime==4.6
appdirs==1.4.3
asn1ate==0.5
backcall==0.1.0
bleach==3.0.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
future==0.16.0
graphviz==0.9
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.0
packaging==16.8
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
py==1.4.33
pyasn1==0.2.3
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.0.7
python-dateutil==2.7.3
pyzmq==17.1.2
qtconsole==4.4.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
